requirements.txt
networkx==3.2.1
lxml==5.1.0
transformers==4.37.2
torch==2.2.0
langchain==0.1.12
python-dotenv==1.0.1
openai==1.14.0
tiktoken==0.6.0
sentence-transformers==2.5.1
numpy==1.26.4 